new updates from the demo project

+++ users/views.py
@@ -16,6 +16,7 @@ from django.contrib.auth import login, logout
 from django.core.mail import send_mail, BadHeaderError
 from django.core.paginator import Paginator,EmptyPage, PageNotAnInteger
 from django.contrib.auth.decorators import user_passes_test
+from django.template.loader import render_to_string
 from django.contrib.auth import login as auth_login, authenticate
 
 class UserPasswordChangeView(LoginRequiredMixin, View):
@@ -33,9 +34,7 @@ class UserPasswordChangeView(LoginRequiredMixin, View):
             messages.success(request, f'Password updated successfully.')
         return render(request, self.template_name, {'form': form})
 
-
 def reset_user_password(request, user_pk):
-    '''Reset user password'''
     user = get_object_or_404(User, pk=user_pk)
     if not request.user.is_authenticated:
         return HttpResponseForbidden()
@@ -46,18 +45,13 @@ def reset_user_password(request, user_pk):
     context = {'mod_user': user,
                'password': password}
     return render(request, 'users/home/reset_user_password.html', context)
-
 @login_required
 def view_profile(request):
     context={}
-    try:
-        member_id=request.user.full_name.id
-    except:
-        member_id=request.user.id
-        
+    member_id=request.user.full_name.id
     pledges=Pledges.objects.filter(Pledge_Made_By_id=member_id)
-    tithes=Revenues.objects.filter(Revenue_filter='tithes',Member_Id=member_id)
-    thanks=Revenues.objects.filter(Revenue_filter='thanks',Member_Id=member_id)
+    tithes=Revenues.objects.filter(Revenue_filter='tithes',Member_Name_id=member_id)
+    thanks=Revenues.objects.filter(Revenue_filter='thanks',Member_Name_id=member_id)
     context['thanks']=thanks
     context['pledges']=pledges
     context['tithes']=tithes
@@ -66,11 +60,7 @@ def view_profile(request):
 
 @login_required
 def edit_profile(request):
-    try:
-        get_member = Members.objects.get(id=request.user.full_name.id)
-    except:
-        get_member = Members.objects.get(id=request.user.id)
-        
+    get_member = Members.objects.get(id=request.user.full_name.id)
     if request.method == 'POST':
         form = MembersForm(request.POST or None, request.FILES or None, instance=get_member)
         if form.is_valid():
@@ -87,18 +77,16 @@ def edit_profile(request):
         args = {'form': form}
         return render(request, 'users/home/update_profile.html', args) 
 
-
 @login_required
 def register(request):
-    '''create an account for user from the dashboard'''
     if request.method == 'POST':
         form = RegisterForm(request.POST)
-        print(form.errors)
         if form.is_valid():
-            form.save()
-            username = form.cleaned_data.get('username')
+            user=form.save()
             messages.success(request, f'Account has been created successfully!, User can now Login')
             return redirect('register')
+        form.errors()
+        return redirect('register')
     else:
         form=RegisterForm()
         users = User.objects.filter(full_name__is_active=True)
@@ -111,11 +99,11 @@ def register(request):
 def MemberAccountRegister(request):
     members=Members.objects.all()
     if request.method == 'POST':
-        form = MembershipAccountForm(request.POST)
+        form = MembershipAccountForm(request.POST, request.FILES)
         if form.is_valid():
             user = form.save()
             login(request, user)
-            member = Members.objects.create(username=user.username, created_by=user,First_Name=user.fname, Second_Name=user.lname, Email=user.email, Photo=user.vatar)
+            member = Members.objects.create(Photo=user.avatar, username=user.username, created_by=user,First_Name=user.fname, Second_Name=user.lname, Email=user.email)
             messages.success(request, f'Account has been created successfully!, Please Complete the registration')
             current_site = get_current_site(request)  
             mail_subject = user.fname + " " + user.lname + ' Created Account'  
@@ -133,33 +121,39 @@ def MemberAccountRegister(request):
         form = MembershipAccountForm()
     return render(request, 'users/home/membershipaccount.html', {'form': form,'members':members})
     
-@login_required    
+   
 def member_profile(request):
-    current_user = request.user.username
-    members_created_by_a_user = Members.objects.filter(username=request.user)
-    try:
-        member = Members.objects.get(created_by=request.user)
-        context={'members':members_created_by_a_user,'member':member, 'current_user':current_user}
-    except:
-        member = Members.objects.get(Full_Named=request.user.full_name)
-        context={'members':members_created_by_a_user, 'member':member, 'current_user':current_user}
-    return render(request,'home/profile.html',context)
+    if not request.user.is_authenticated:
+        return redirect('login')
+    else:
+        current_user = request.user.username
+        members_created_by_a_user = Members.objects.filter(username=request.user)
+        try:
+            member = Members.objects.get(created_by=request.user)
+            context={'members':members_created_by_a_user,'member':member, 'current_user':current_user}
+        except:
+            member = Members.objects.get(Full_Named=request.user.full_name)
+            context={'members':members_created_by_a_user, 'member':member, 'current_user':current_user}
+        return render(request,'home/profile.html',context)
 
+@login_required
 def delete_user(request,pk):
     user= get_object_or_404(User, id=pk)
-    user.delete()
-    messages.success(request, "The User successfully deleted!")
-    return redirect("register")
+    if request.method == "GET":
+        user.delete()
+        messages.success(request, "The User successfully deleted!")
+        return redirect("register")
+    context= {'user': user}
+    return render(request, 'users/home/delete_user.html', context)
     
- 
+    
 @login_required
-def update_system_user(request, user_pk):
+def user_updated(request, user_pk):
     user = get_object_or_404(User, pk=user_pk)
     if request.method == "POST":
         form = RegisterForm(request.POST, request.FILES, instance=user)
         if form.is_valid():
             form = form.save()
-            messages.success(request, "The User info successfully updated!")
             return redirect('register')
         else:
             form = RegisterForm(instance=user)
@@ -170,8 +164,9 @@ def update_system_user(request, user_pk):
         args = {'form': form,}
         return render(request, 'users/home/user_update.html', args)
 
+
 @login_required
-def church_user_account(request, user_pk):
+def user_update(request, user_pk):
     user = get_object_or_404(Members, created_by=user_pk)
     if request.method == "POST":
         form = MembersForm(request.POST, request.FILES, instance=user)
@@ -192,9 +187,11 @@ def logout_request(request):
     logout(request)
     messages.info(request, "Logged out successfully!")
     return HttpResponseRedirect(request.GET.get('next','/'))
-
+    
 def UserLogin(request):
     form = UserLoginForm()
+    redirect_to = request.GET.get('next', '')
+    current_url = request.path
     if request.method == 'POST':
         form = UserLoginForm(request.POST)
         if form.is_valid():
@@ -203,10 +200,14 @@ def UserLogin(request):
             user = authenticate(username=username, password=password)
             if user is not None:
                 auth_login(request, user)
-                current_url = request.path
-                print(current_url)
-                return redirect('member_profile')
-            
+                
+                if current_url == '/Login/' and not redirect_to:
+                    return redirect('member_profile')
+                else:
+                    return HttpResponseRedirect(redirect_to) 
+                    
+            else:
+                return render(request, 'users/login.html', {'form':form})
         else:
             return render(request, 'users/login.html', {'form':form})
